move utils to pkg
