+++ pkg/cmd/find.go
@@ -1,0 +1,16 @@
+package cmd
+
+import "runtime"
+
+var (
+	platform = runtime.GOOS
+)
+
+func getFindCommand() string {
+	switch platform {
+	case "windows":
+		return "where"
+	default:
+		return "which"
+	}
+}
